pyproject.toml
[tool.poetry]
name = "balapan"
version = "0.1.0"
description = ""
authors = ["Anas Kaarov <anasbinkurban@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.25.1"
django = "^4.1.6"
django-environ = "^0.9.0"
gunicorn = "^20.1.0"
django-4-jet = "^1.0.9"
pillow = "^9.4.0"
django-solo = "^2.0.0"
django-filter = "^22.1"
django-cors-headers = "^3.14.0"
django-silk = "^5.0.3"
requests = "^2.31.0"
six = "^1.16.0"
aiogram = "2.25.1"
django-rest-swagger = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
